+++ src/main/java/com/nekomaster1000/infernalexp/entities/GlowsquitoEntity.java
@@ -236,7 +236,7 @@ public class GlowsquitoEntity extends AnimalEntity implements IFlyingAnimal {
             vector3d = GlowsquitoEntity.this.getLook(0.0F);
 
             Vector3d vector3d2 = RandomPositionGenerator.findAirTarget(GlowsquitoEntity.this, 8, 7, vector3d, ((float)Math.PI / 2F), 2, 1);
-            return vector3d2 != null ? vector3d2 : RandomPositionGenerator.findGroundTarget(GlowsquitoEntity.this, 8, 4, -2, vector3d, (double)((float)Math.PI / 2F));
+            return vector3d2 != null ? vector3d2 : RandomPositionGenerator.findGroundTarget(GlowsquitoEntity.this, 8, 4, -2, vector3d, (float)Math.PI / 2F);
         }
     }
